// src/types.ts
export type TagCategory =
  | "fandom"
  | "character"
  | "relationship"
  | "archive warning"
  | "additional tags";

export interface Tag {
  name: string;
  // Not all tags have user-facing IDs Example: additional tags.
  // TODO: figure out other types (or whether they can be extracted from somewhere else).
  id: string | null;
  category: TagCategory;
  canonical: boolean;
  common: boolean;
  // Canonical name will be the same as "name" on canonical tags, and different on tags
  // that have been synned to a canonical. It will be null when tags haven't been marked as
  // common and cannot be filtered on.
  canonicalName: string | null;
}

export interface WorkData {
  workId: string;
  chapterId?: string;
  collectionName?: string;
}

export interface User {
  id: string;
  name: string;
  pseuds: string;
  url: string;
  joined: string;
  location: string | null;
  email: string | null;
  birthday: string | null;
  bioHtml: string | null;
}